feat: verify dynamic configuration
